+++ packages/adapters/eth/src/websocket/metrics.rs
@@ -1,43 +1,7 @@
-use std::{
-    cmp::min, num::NonZeroU32, ops::RangeInclusive, str::FromStr, sync::Arc, time::Duration,
-};
 
-use crate::estimation::{MaxTxFeesPerGas, TransactionRequestExt};
-use ::metrics::RegistersMetrics;
-use alloy::{
-    consensus::{SignableTransaction, Transaction},
-    eips::{
-        BlockNumberOrTag,
-        eip4844::{BYTES_PER_BLOB, DATA_GAS_PER_BLOB},
-    },
-    network::{Ethereum, EthereumWallet, TransactionBuilder, TransactionBuilder4844, TxSigner},
-    primitives::{Address, B256, ChainId},
-    providers::{
-        Provider, ProviderBuilder, SendableTx, WsConnect,
-        utils::{EIP1559_FEE_ESTIMATION_PAST_BLOCKS, Eip1559Estimation},
-    },
-    pubsub::PubSubFrontend,
-    rpc::types::{FeeHistory, TransactionReceipt, TransactionRequest},
-    signers::{Signature, local::PrivateKeySigner},
-    sol,
-};
-use itertools::Itertools;
+use alloy::eips::eip4844::BYTES_PER_BLOB;
 use metrics::prometheus::{self, histogram_opts};
-use serde::Deserialize;
-use services::{
-    state_committer::port::l1::Priority,
-    types::{
-        BlockSubmissionTx, Fragment, FragmentsSubmitted, L1Tx, NonEmpty, TransactionResponse, U256,
-    },
-};
-use tracing::info;
-use url::Url;
 
-use crate::{
-    AwsClient, AwsConfig, Error, Result, blob_encoder,
-    failover_client::{ProviderConfig, ProviderInit},
-    provider::L1Provider,
-};
 #[derive(Clone)]
 pub struct Metrics {
     pub(crate) blobs_per_tx: prometheus::Histogram,
